[#11] refactor: 게시판 회원 기능 추가

+++ src/main/java/kr/soulware/crudsoulware/posts/service/PostsService.java
@@ -2,11 +2,15 @@ package kr.soulware.crudsoulware.posts.service;
 
 import kr.soulware.crudsoulware.exception.ErrorCode;
 import kr.soulware.crudsoulware.exception.model.NotFoundException;
+import kr.soulware.crudsoulware.exception.model.UnauthorizedException;
 import kr.soulware.crudsoulware.posts.dto.request.PostsSaveRequestDto;
 import kr.soulware.crudsoulware.posts.dto.request.PostsUpdateRequestDto;
 import kr.soulware.crudsoulware.posts.dto.response.PostResponseDto;
 import kr.soulware.crudsoulware.posts.entity.Posts;
 import kr.soulware.crudsoulware.posts.repository.PostsRepository;
+import kr.soulware.crudsoulware.security.UserDetailsImpl;
+import kr.soulware.crudsoulware.user.entity.User;
+import kr.soulware.crudsoulware.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,16 +22,27 @@ import java.util.List;
 public class PostsService {
 
     private final PostsRepository postsRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public Long save(PostsSaveRequestDto requestDto) {
-        Posts posts = requestDto.toEntity();
+    public Long save(PostsSaveRequestDto requestDto, UserDetailsImpl loginUser) {
+        Posts posts = generatePosts(requestDto, loginUser);
         return postsRepository.save(posts).getId();
     }
 
+    private Posts generatePosts(PostsSaveRequestDto requestDto, UserDetailsImpl loginUser) {
+        User user = findUserByEmail(loginUser.getEmail());
+        return Posts.builder()
+                .title(requestDto.getTitle())
+                .content(requestDto.getContent())
+                .author(user)
+                .build();
+    }
+
     @Transactional
-    public Long update(Long id, PostsUpdateRequestDto requestDto) {
+    public Long update(Long id, PostsUpdateRequestDto requestDto, UserDetailsImpl loginUser) {
         Posts posts = findPostsById(id);
+        checkAuthor(posts, loginUser);
         posts.update(requestDto.getTitle(), requestDto.getContent());
         return id;
     }
@@ -45,8 +60,9 @@ public class PostsService {
     }
 
     @Transactional
-    public Long delete(Long id) {
+    public Long delete(Long id, UserDetailsImpl loginUser) {
         Posts posts = findPostsById(id);
+        checkAuthor(posts, loginUser);
         postsRepository.delete(posts);
         return id;
     }
@@ -55,6 +71,24 @@ public class PostsService {
         return postsRepository.findById(postsId)
                 .orElseThrow(() -> new NotFoundException(
                         ErrorCode.NOT_FOUND_POST_EXCEPTION,
-                        ErrorCode.NOT_FOUND_POST_EXCEPTION.getMessage()));
+                        ErrorCode.NOT_FOUND_POST_EXCEPTION.getMessage()
+                ));
+    }
+
+    private User findUserByEmail(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new NotFoundException(
+                        ErrorCode.NOT_FOUND_USER_EXCEPTION,
+                        ErrorCode.NOT_FOUND_USER_EXCEPTION.getMessage()
+                ));
+    }
+
+    private void checkAuthor(Posts posts, UserDetailsImpl loginUser) {
+        if (!posts.getAuthor().getEmail().equals(loginUser.getEmail())) {
+            throw new UnauthorizedException(
+                    ErrorCode.INSUFFICIENT_AUTHENTICATION_TO_POSTS,
+                    ErrorCode.INSUFFICIENT_AUTHENTICATION_TO_POSTS.getMessage()
+            );
+        }
     }
 }
